Update api.yaml
should fix the feedback Do not simulate a procedure call in APIs. REST dictates that you must reason in terms of collections and items. For example, putting a like should be designed as adding a like to the collection of likes. The same when deleting an item from a collection.